chore: moved core constants to utils
