+++ python/sauron-core/core/utils/constants.py
